Add a test exercising a veneur forwarding chain
This is as end-to-end as I think we can make it: The test sets up a
local veneur, a veneur proxy, and a global veneur, each forwarding
along to the next. Then the local veneur ingests an indicator span and
flushes it all the way across, and checks that the global veneur is
correctly reporting the proper indicator timer histogram.